Version 0.9
+ Split up the code into different header and source code files.

  New header files: Printing.h, Search.h, Sorting.h,
  SongManagementStruct.h

  New source code files: Printing.cpp, Search.cpp, Sorting.cpp

+ Changed main.cpp file according to the changes

+ Minor output fixes
