+++ Search.h
@@ -1,0 +1,8 @@
+#include "SongManagementStruct.h"
+
+#ifndef SEARCH_H
+#define SEARCH_H
+
+void search(std::string searchName, const int NUMBER_OF_SONGS, SongManagement table[], const char **targetName[], bool found);
+
+#endif
