+++ SongManagementStruct.h
@@ -1,0 +1,12 @@
+#ifndef STRUCT_H
+#define STRUCT_H
+
+struct SongManagement
+{
+    const char *track;
+    const char *band;
+    int popularity;
+};
+
+//extern SongManagement glovalStruct;
+#endif
